requirements.txt
flask==2.3.3
flask-socketio==5.3.4
pycryptodome==3.18.0
python-dotenv==1.0.0
cryptography==41.0.3
colorama==0.4.6
python-socketio==5.7.2
websocket-client==1.8.0 